make seq into unsigned long long

+++ msg_info.h
@@ -2,7 +2,7 @@
 #define _MSG_INFO_H  1
 
 struct MSG_INFO{
-	  int seqnumber;
+	  unsigned long long seqnumber;
 	  double time_stamp;
 	  //double delay_type;
 	  int payload_size;
